Generate Random Given and Surnames

Functions for generating random gender and ethnicity correct first and/or last names where names are proportionally sampled based upon their frequency in a large scale database.

Details

Package:

randomNames

Type:

Package

Version:

1.5-0.0

Date:

2021-4-21

License:

GPL-3

LazyLoad:

yes