Function for Generating Random Names and a Dataset

Functions for generating random gender and ethnicity correct first and/or last names where names are proportionally sampled based upon their frequency in a large scale database.

Details

Package:

randomNames

Type:

Package

Version:

1.0-0.0

Date:

2017-6-5

License:

GPL-3

LazyLoad:

yes